feat(assignment): extract assignment
